Create compare-pdf utility to compare pdfs.
This utility converts each page of the pdf to a png, and then carries out two checks
to determine whether the pdfs are identical:

* a LengthComparison, which checks that the documents have equal lengths
* an ImageComparison, which checks that the difference between the images (calculated as a root mean square error)
  falls below the specified threshold (default: 10).

If any of these checks fail, the comparison fails and the pdfs are considered not to be similar.

The entrypoint is PdfComparer(...).compare(). This function returns a tuple containing a boolean (the result of the comparison,
and a reasons array, containing the reasons why the comparison failed.